ReportGenerator/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
celery==5.3.6
jinja2==3.1.3
reportlab==4.1.0
matplotlib==3.8.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
boto3==1.34.45
python-multipart==0.0.9
pydantic==2.6.2
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
